Refactored method and moved to factory class to incorporate code review comment. Fixed lint error by removing the type of port. and other minor changes.
