+++ sdk/python/feast/permissions/client/http_auth_requests_wrapper.py
@@ -5,7 +5,7 @@ from feast.permissions.auth.auth_type import AuthType
 from feast.permissions.auth_model import (
     AuthConfig,
 )
-from feast.permissions.client.auth_client_manager import get_auth_client_manager
+from feast.permissions.client.auth_client_manager_factory import get_auth_token
 
 
 class AuthenticatedRequestsSession(Session):
@@ -18,6 +18,5 @@ def get_http_auth_requests_session(auth_config: AuthConfig) -> Session:
     if auth_config.type == AuthType.NONE.value:
         request_session = requests.session()
     else:
-        auth_client_manager = get_auth_client_manager(auth_config)
-        request_session = AuthenticatedRequestsSession(auth_client_manager.get_token())
+        request_session = AuthenticatedRequestsSession(get_auth_token(auth_config))
     return request_session
